project_management/requirements.txt
Django==5.0.6
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
django-filter==24.2
drf-yasg==1.21.7
gunicorn>=20.1,<21.0
psycopg2-binary==2.9.9
django-cors-headers==4.3.1